changed decision on remote or local
- loginpresenter decides when to go remote or local
- remote crud operations impl refactored

+++ app/src/main/java/com/example/avenger/mad2017retry/database/RemoteCRUDOperationsImpl.java
@@ -1,5 +1,6 @@
 package com.example.avenger.mad2017retry.database;
 
+import android.content.Context;
 import android.os.AsyncTask;
 import android.util.Log;
 
@@ -20,10 +21,6 @@ import retrofit2.http.Path;
 
 
 public class RemoteCRUDOperationsImpl implements ICRUDOperationsAsync {
-
-    private static String WEB_API_BASE_URL = "http://localhost:8080/";
-    protected static String logger = RemoteCRUDOperationsImpl.class.getSimpleName();
-
     //inner interface to define webapi calls to given remote database application
     public interface ICRUDWebApi {
         @POST("/api/todos")
@@ -40,17 +37,24 @@ public class RemoteCRUDOperationsImpl implements ICRUDOperationsAsync {
 
         @DELETE("/api/todos/{id}")
         Call<Boolean> deleteToDoItem(@Path("id") long id);
+
+        //TODO implement clearDB method
     }
 
+    private static String WEB_API_BASE_URL = "http://localhost:8080/";
+    protected static String logger = RemoteCRUDOperationsImpl.class.getSimpleName();
+
+    private LocalCRUDOperationsImpl localCRUD;
     private ICRUDWebApi webAPI;
 
-    public RemoteCRUDOperationsImpl() {
+    public RemoteCRUDOperationsImpl(Context context) {
         Retrofit retrofit = new Retrofit.Builder()
                 .baseUrl(WEB_API_BASE_URL)
                 .addConverterFactory(GsonConverterFactory.create())
                 .build();
 
         this.webAPI = retrofit.create(ICRUDWebApi.class);
+        this.localCRUD = new LocalCRUDOperationsImpl(context);
     }
 
 
@@ -59,8 +63,10 @@ public class RemoteCRUDOperationsImpl implements ICRUDOperationsAsync {
         new AsyncTask<Todo, Void, Todo>() {
             @Override
             protected Todo doInBackground(Todo... params) {
+                final Todo[] newItem = {new Todo("", "")};
+                localCRUD.createToDo(item, result -> newItem[0] = result);
                 try {
-                    return webAPI.createToDoItem(params[0]).execute().body();
+                    return webAPI.createToDoItem(newItem[0]).execute().body();
                 } catch (IOException e) {
                     Log.d(logger, "Item not created.");
                     e.printStackTrace();
@@ -99,7 +105,6 @@ public class RemoteCRUDOperationsImpl implements ICRUDOperationsAsync {
     @Override
     public void readToDo(long id, final CallbackFunction<Todo> callback) {
         new AsyncTask<Long, Void, Todo>() {
-
             @Override
             protected Todo doInBackground(Long... params) {
                 try {
@@ -121,11 +126,12 @@ public class RemoteCRUDOperationsImpl implements ICRUDOperationsAsync {
     @Override
     public void updateToDo(long id, final Todo item, final CallbackFunction<Todo> callback) {
         new AsyncTask<Long, Void, Todo>() {
-
             @Override
             protected Todo doInBackground(Long... params) {
+                final Todo[] newItem = {new Todo("", "")};
+                localCRUD.updateToDo(id, item, result -> newItem[0] = result);
                 try {
-                    return webAPI.updateToDoItem(params[0], item).execute().body();
+                    return webAPI.updateToDoItem( newItem[0].getId(),  newItem[0]).execute().body();
                 } catch (IOException e) {
                     Log.d(logger, "Could not update item with id: " + params[0]);
                     e.printStackTrace();
@@ -145,12 +151,18 @@ public class RemoteCRUDOperationsImpl implements ICRUDOperationsAsync {
         new AsyncTask<Long, Void, Boolean>() {
             @Override
             protected Boolean doInBackground(Long... params) {
-                try {
-                    return webAPI.deleteToDoItem(params[0]).execute().body();
-                } catch (IOException e) {
-                    Log.d(logger, "Could not delete item with id: " + params[0]);
-                    e.printStackTrace();
+                final boolean[] deletedLocally = {false};
+                localCRUD.deleteToDo(id, result -> deletedLocally[0] = result);
+                if (deletedLocally[0]) {
+                    try {
+                        return webAPI.deleteToDoItem(params[0]).execute().body();
+                    } catch (IOException e) {
+                        Log.d(logger, "Could not delete item with id: " + params[0]);
+                        e.printStackTrace();
+                    }
+                    return false;
                 }
+                Log.d(logger, "Could not delete todo locally.");
                 return false;
             }
 
@@ -159,6 +171,5 @@ public class RemoteCRUDOperationsImpl implements ICRUDOperationsAsync {
                 callback.process(aBoolean);
             }
         }.execute(id);
-
     }
 }
